feat: update doc for running Ingress API conformance tests
The commit adds support for running Ingress API conformance tests. The tests are defined in the `test/e2e/conformance/tests` directory and are registered in the `ConformanceTests` global variable. This allows developers to select and run specific conformance tests. The commit also updates the documentation to reflect the changes.

Co-authored-by: Kent Dong <ch3cho@qq.com>